requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
colorama==0.4.3
# Django==3.0.5
Django==2.0.5 
django-debug-toolbar==2.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.1
typed-ast==1.4.0
wrapt==1.11.2
Pillow==5.1.0
